chore: move a inicialização do servidor para um arquivo a parte
